refactor(opevents,apirouter): typed constructor for tenant.subscription.updated
Moves the payload struct next to the other operator-event payloads and
replaces apirouter's hand-built event (raw topic string) with the
constructor, so the topic constant and tenant pairing live in one place.
